@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 29
     buildToolsVersion "27.1.3"
     defaultConfig {
         applicationId "com.zby.wheeldemo"
         minSdkVersion 15
-        targetSdkVersion 27
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -20,8 +20,8 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
 
     implementation project(path:':wheelview')
-//    implementation 'com.github.biby0915:WheelView:1.0.3'
+//    implementation 'com.github.biby0915:WheelView:1.0.2'
 }
